+++ src/Welcome.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+export default function Welcome() {
+  return (
+    <div className="main">
+      <h1>Rock, Paper, Scissors</h1>
+      <p>Do you dare to play</p>
+      <p>this game?</p>
+      <button>Play</button>
+    </div>
+  );
+}
